# app/static/swagger.yaml
openapi: 3.0.0
info:
  title: Bus Payments & Rewards API
  description: Comprehensive API for bus payment system with offers, rewards, bookings, and refunds
  version: 1.0.0
servers:
  - url: http://localhost:{port}
    description: Local development server
    variables:
      port:
        default: '5006'
        description: Server port number
...
paths:
...
  /api/bookings/{booking_id}/status:
    get:
      tags:
        - Bookings
      summary: Get booking status
      parameters:
        - in: path
          name: booking_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Booking status details
        '404':
          description: Booking not found
        '500':
          description: Internal server error
  /api/bookings/{booking_id}/modify:
    put:
      tags:
        - Bookings
      summary: Modify booking
      parameters:
        - in: path
          name: booking_id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                new_booking_date:
                  type: string
                  format: date-time
                modification_reason:
                  type: string
                additional_services:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Booking modified successfully
        '400':
          description: Invalid modification request
        '404':
          description: Booking not found
        '500':
          description: Internal server error

  # Integration & Simulator Endpoints
...